Move towards making this a stub package on 0.7
Now that showarg has moved to Base, this package isn't goint to be necessary. This tries to make it as compatible as possible, but unfortunately it seems that packages are going to have to handle most of the version-checking directly.

Because functionality will be moving to individual packages, for the tests here it's safer to build a custom array type than to rely on a package
